# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Film API Documentation",
        "description": "API documentation for Film management",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/typen": {
            "get": {
                "tags": [
                    "Typen"
                ],
                "summary": "Liste aller Tickettypen",
                "operationId": "55de1face28ab22b3c253491bc3c3cc6",
                "responses": {
                    "200": {
                        "description": "Liste von Tickettypen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Typ"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Typen"
                ],
                "summary": "Erstelle einen neuen Tickettyp",
                "operationId": "68a69ef718bbf04b4e051a6ff0f51d46",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tickettyp erfolgreich erstellt",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Typ"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/typen/{id}": {
            "get": {
                "tags": [
                    "Typen"
                ],
                "summary": "Zeige einen Tickettyp",
                "operationId": "74e505f4fbe3bfc5aa32585d25262311",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Tickettyps",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tickettyp",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Typ"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tickettyp nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Typ nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Typen"
                ],
                "summary": "Aktualisiere einen Tickettyp",
                "operationId": "52391048ad320dddc01ee6696313185e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Tickettyps",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TypInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tickettyp aktualisiert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Typ"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tickettyp nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Typ nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Typen"
                ],
                "summary": "Lösche einen Tickettyp",
                "operationId": "ce5670b45ecb253c5ed79b10d7e26945",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID des Tickettyps",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Tickettyp gelöscht"
                    },
                    "404": {
                        "description": "Tickettyp nicht gefunden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Typ nicht gefunden."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "resource_not_found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Tickettyp wird noch verwendet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Typ kann nicht gelöscht werden, da er von einem oder mehreren Besuchen verwendet wird."
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "foreign_key_constraint_violation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Typ": {
                "required": [
                    "id",
                    "name",
                    "zuschlag_chf"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "VIP"
                    },
                    "zuschlag_chf": {
                        "type": "number",
                        "format": "float",
                        "example": 5.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TypInput": {
                "required": [
                    "name",
                    "zuschlag_chf"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "VIP"
                    },
                    "zuschlag_chf": {
                        "type": "number",
                        "format": "float",
                        "example": 5.5
                    }
                },
                "type": "object"
            }
        }
    },
...
}